Split presentation module
