+++ .idea/artifacts/api.xml
@@ -1,8 +1,8 @@
 <component name="ArtifactManager">
   <artifact type="jar" name="api">
-    <output-path>$PROJECT_DIR$/domain/api/build/libs</output-path>
+    <output-path>$PROJECT_DIR$/presentation/api/build/libs</output-path>
     <root id="archive" name="api.jar">
-      <element id="module-output" name="Valiutchik.domain.api.androidMain" />
+      <element id="module-output" name="Valiutchik.presentation.api.androidMain" />
     </root>
   </artifact>
 </component>
