+++ .idea/artifacts/api_desktop.xml
@@ -1,8 +1,8 @@
 <component name="ArtifactManager">
   <artifact type="jar" name="api-desktop">
-    <output-path>$PROJECT_DIR$/domain/api/build/libs</output-path>
+    <output-path>$PROJECT_DIR$/presentation/api/build/libs</output-path>
     <root id="archive" name="api-desktop.jar">
-      <element id="module-output" name="Valiutchik.domain.api.desktopMain" />
+      <element id="module-output" name="Valiutchik.presentation.api.desktopMain" />
     </root>
   </artifact>
 </component>
