+++ .idea/artifacts/api_wasm_js.xml
@@ -1,8 +1,8 @@
 <component name="ArtifactManager">
   <artifact type="jar" name="api-wasm-js">
-    <output-path>$PROJECT_DIR$/domain/api/build/libs</output-path>
+    <output-path>$PROJECT_DIR$/presentation/api/build/libs</output-path>
     <root id="archive" name="api-wasm-js.jar">
-      <element id="module-output" name="Valiutchik.domain.api.wasmJsMain" />
+      <element id="module-output" name="Valiutchik.presentation.api.wasmJsMain" />
     </root>
   </artifact>
 </component>
