+++ .idea/artifacts/impl.xml
@@ -1,8 +1,8 @@
 <component name="ArtifactManager">
   <artifact type="jar" name="impl">
-    <output-path>$PROJECT_DIR$/domain/impl/build/libs</output-path>
+    <output-path>$PROJECT_DIR$/presentation/impl/build/libs</output-path>
     <root id="archive" name="impl.jar">
-      <element id="module-output" name="Valiutchik.domain.impl.androidMain" />
+      <element id="module-output" name="Valiutchik.presentation.impl.androidMain" />
     </root>
   </artifact>
 </component>
