+++ .idea/artifacts/impl_desktop.xml
@@ -1,8 +1,8 @@
 <component name="ArtifactManager">
   <artifact type="jar" name="impl-desktop">
-    <output-path>$PROJECT_DIR$/domain/impl/build/libs</output-path>
+    <output-path>$PROJECT_DIR$/presentation/impl/build/libs</output-path>
     <root id="archive" name="impl-desktop.jar">
-      <element id="module-output" name="Valiutchik.domain.impl.desktopMain" />
+      <element id="module-output" name="Valiutchik.presentation.impl.desktopMain" />
     </root>
   </artifact>
 </component>
