+++ .idea/artifacts/impl_wasm_js.xml
@@ -1,8 +1,8 @@
 <component name="ArtifactManager">
   <artifact type="jar" name="impl-wasm-js">
-    <output-path>$PROJECT_DIR$/domain/impl/build/libs</output-path>
+    <output-path>$PROJECT_DIR$/presentation/impl/build/libs</output-path>
     <root id="archive" name="impl-wasm-js.jar">
-      <element id="module-output" name="Valiutchik.domain.impl.wasmJsMain" />
+      <element id="module-output" name="Valiutchik.presentation.impl.wasmJsMain" />
     </root>
   </artifact>
 </component>
