+++ .idea/artifacts/wiring.xml
@@ -1,8 +1,8 @@
 <component name="ArtifactManager">
   <artifact type="jar" name="wiring">
-    <output-path>$PROJECT_DIR$/domain/wiring/build/libs</output-path>
+    <output-path>$PROJECT_DIR$/presentation/wiring/build/libs</output-path>
     <root id="archive" name="wiring.jar">
-      <element id="module-output" name="Valiutchik.domain.wiring.androidMain" />
+      <element id="module-output" name="Valiutchik.presentation.wiring.androidMain" />
     </root>
   </artifact>
 </component>
