+++ .idea/artifacts/wiring_desktop.xml
@@ -1,8 +1,8 @@
 <component name="ArtifactManager">
   <artifact type="jar" name="wiring-desktop">
-    <output-path>$PROJECT_DIR$/domain/wiring/build/libs</output-path>
+    <output-path>$PROJECT_DIR$/presentation/wiring/build/libs</output-path>
     <root id="archive" name="wiring-desktop.jar">
-      <element id="module-output" name="Valiutchik.domain.wiring.desktopMain" />
+      <element id="module-output" name="Valiutchik.presentation.wiring.desktopMain" />
     </root>
   </artifact>
 </component>
