+++ .idea/artifacts/wiring_wasm_js.xml
@@ -1,8 +1,8 @@
 <component name="ArtifactManager">
   <artifact type="jar" name="wiring-wasm-js">
-    <output-path>$PROJECT_DIR$/domain/wiring/build/libs</output-path>
+    <output-path>$PROJECT_DIR$/presentation/wiring/build/libs</output-path>
     <root id="archive" name="wiring-wasm-js.jar">
-      <element id="module-output" name="Valiutchik.domain.wiring.wasmJsMain" />
+      <element id="module-output" name="Valiutchik.presentation.wiring.wasmJsMain" />
     </root>
   </artifact>
 </component>
