+++ desktop/build.gradle.kts
@@ -39,7 +39,7 @@ kotlin {
                 implementation(project(":api:wiring"))
                 implementation(project(":data:wiring"))
                 implementation(project(":domain:wiring"))
-                implementation(project(":presentation"))
+                implementation(project(":presentation:wiring"))
                 implementation(libs.compose.ui)
                 implementation(libs.compose.resources)
                 implementation(libs.compose.material)
