+++ presentation/api/build.gradle.kts
@@ -1,0 +1,51 @@
+/*
+ *    Copyright 2026 Andrey Mukamolov
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+@file:OptIn(ExperimentalWasmDsl::class)
+
+import org.jetbrains.kotlin.gradle.ExperimentalWasmDsl
+
+plugins {
+    id("buildlogic.library-conventions")
+    alias(libs.plugins.android.lint)
+}
+
+kotlin {
+    android {
+        namespace = "dev.fobo66.valiutchik.presentation"
+    }
+
+    wasmJs {
+        browser {
+            testTask {
+                enabled = false
+            }
+        }
+    }
+
+    sourceSets {
+        commonMain {
+            dependencies {
+                implementation(project(":domain:api"))
+                implementation(libs.kotlinx.coroutines.core)
+                implementation(libs.kotlinx.collections)
+                implementation(libs.kotlinx.datetime)
+                implementation(libs.androidx.lifecycle.viewmodel)
+                implementation(libs.compose.stable.marker)
+            }
+        }
+    }
+}
