+++ presentation/api/src/commonMain/kotlin/dev/fobo66/valiutchik/presentation/entity/MainScreenStateTrigger.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2025 Andrey Mukamolov
+ *    Copyright 2026 Andrey Mukamolov
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
 
 package dev.fobo66.valiutchik.presentation.entity
 
-internal data class MainScreenStateTrigger(
+data class MainScreenStateTrigger(
     val isRefreshTriggered: Boolean,
     val isLocationAvailable: Boolean?
 )
