+++ settings.gradle.kts
@@ -64,8 +64,10 @@ include(
     ":api:api",
     ":api:impl",
     ":api:wiring",
+    ":presentation:api",
+    ":presentation:impl",
+    ":presentation:wiring",
     ":baselineprofile",
-    ":presentation",
     ":widget",
     ":ui",
     ":desktop",
